+++ packages/e2e-tests/import-queries/svelte.config.js
@@ -1,0 +1,10 @@
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
+
+export default {
+	preprocess: [vitePreprocess()],
+	onwarn(warning, defaultHandler) {
+		// import query test generates one of these
+		if (warning.code === 'custom-element-no-tag') return;
+		defaultHandler(warning);
+	}
+};