+++ packages/e2e-tests/preprocess-with-vite/src/Foo.svelte
@@ -1,6 +1,8 @@
 <script lang="ts">
 	let foo: string = 'stylus';
 	export let bla: string = 'blub';
+	console.log('pure test 1', new Date());
+	console.log('pure test 2');
 </script>
 
 <h1 id="foo-title">Styles with {foo} {bla}</h1>