+++ packages/playground/big/vite.config.js
@@ -1,12 +1,6 @@
-const { svelte } = require('@sveltejs/vite-plugin-svelte');
-const { defineConfig } = require('vite');
+import { svelte } from '@sveltejs/vite-plugin-svelte';
+import { defineConfig } from 'vite';
 
-module.exports = defineConfig(({ command, mode }) => {
-	const isProduction = mode === 'production';
-	return {
-		plugins: [svelte()],
-		build: {
-			minify: isProduction
-		}
-	};
+export default defineConfig({
+	plugins: [svelte()]
 });