+++ packages/vite-plugin-svelte/src/index.ts
@@ -4,8 +4,8 @@ import { HmrContext, ModuleNode, Plugin, ResolvedConfig, UserConfig } from 'vite
 import { isDepExcluded } from 'vitefu';
 import { handleHotUpdate } from './handle-hot-update';
 import { log, logCompilerWarnings } from './utils/log';
-import { CompileData, createCompileSvelte } from './utils/compile';
-import { buildIdParser, IdParser, SvelteRequest } from './utils/id';
+import { type CompileSvelte, createCompileSvelte } from './utils/compile';
+import { buildIdParser, IdParser } from './utils/id';
 import {
 	buildExtraViteConfig,
 	validateInlineOptions,
@@ -23,6 +23,7 @@ import { toRollupError } from './utils/error';
 import { saveSvelteMetadata } from './utils/optimizer';
 import { svelteInspector } from './ui/inspector/plugin';
 import { VitePluginSvelteCache } from './utils/vite-plugin-svelte-cache';
+import { loadRaw } from './utils/load-raw';
 import { FAQ_LINK_DEPRECATED_SVELTE_FIELD } from './utils/constants';
 
 interface PluginAPI {
@@ -46,11 +47,7 @@ export function svelte(inlineOptions?: Partial<Options>): Plugin[] {
 	let options: ResolvedOptions;
 	let viteConfig: ResolvedConfig;
 	/* eslint-disable no-unused-vars */
-	let compileSvelte: (
-		svelteRequest: SvelteRequest,
-		code: string,
-		options: Partial<ResolvedOptions>
-	) => Promise<CompileData>;
+	let compileSvelte: CompileSvelte;
 	/* eslint-enable no-unused-vars */
 
 	let resolvedSvelteSSR: Promise<PartialResolvedId | null>;
@@ -93,7 +90,6 @@ export function svelte(inlineOptions?: Partial<Options>): Plugin[] {
 				if (isSvelteMetadataChanged) {
 					// Force Vite to optimize again. Although we mutate the config here, it works because
 					// Vite's optimizer runs after `buildStart()`.
-					// TODO: verify this works in vite3
 					viteConfig.optimizeDeps.force = true;
 				}
 			},
@@ -104,23 +100,26 @@ export function svelte(inlineOptions?: Partial<Options>): Plugin[] {
 				setupWatchers(options, cache, requestParser);
 			},
 
-			load(id, opts) {
+			async load(id, opts) {
 				const ssr = !!opts?.ssr;
 				const svelteRequest = requestParser(id, !!ssr);
 				if (svelteRequest) {
-					const { filename, query } = svelteRequest;
-					// virtual css module
-					if (query.svelte && query.type === 'style') {
-						const css = cache.getCSS(svelteRequest);
-						if (css) {
-							log.debug(`load returns css for ${filename}`);
-							return css;
+					const { filename, query, raw } = svelteRequest;
+					if (raw) {
+						return loadRaw(svelteRequest, compileSvelte, options);
+					} else {
+						if (query.svelte && query.type === 'style') {
+							const css = cache.getCSS(svelteRequest);
+							if (css) {
+								log.debug(`load returns css for ${filename}`);
+								return css;
+							}
 						}
-					}
-					// prevent vite asset plugin from loading files as url that should be compiled in transform
-					if (viteConfig.assetsInclude(filename)) {
-						log.debug(`load returns raw content for ${filename}`);
-						return fs.readFileSync(filename, 'utf-8');
+						// prevent vite asset plugin from loading files as url that should be compiled in transform
+						if (viteConfig.assetsInclude(filename)) {
+							log.debug(`load returns raw content for ${filename}`);
+							return fs.readFileSync(filename, 'utf-8');
+						}
 					}
 				}
 			},
@@ -129,14 +128,12 @@ export function svelte(inlineOptions?: Partial<Options>): Plugin[] {
 				const ssr = !!opts?.ssr;
 				const svelteRequest = requestParser(importee, ssr);
 				if (svelteRequest?.query.svelte) {
-					if (svelteRequest.query.type === 'style') {
+					if (svelteRequest.query.type === 'style' && !svelteRequest.raw) {
 						// return cssId with root prefix so postcss pipeline of vite finds the directory correctly
 						// see https://github.com/sveltejs/vite-plugin-svelte/issues/14
 						log.debug(`resolveId resolved virtual css module ${svelteRequest.cssId}`);
 						return svelteRequest.cssId;
 					}
-					log.debug(`resolveId resolved ${importee}`);
-					return importee; // query with svelte tag, an id we generated, no need for further analysis
 				}
 
 				if (ssr && importee === 'svelte') {
@@ -211,7 +208,7 @@ export function svelte(inlineOptions?: Partial<Options>): Plugin[] {
 			async transform(code, id, opts) {
 				const ssr = !!opts?.ssr;
 				const svelteRequest = requestParser(id, ssr);
-				if (!svelteRequest || svelteRequest.query.svelte) {
+				if (!svelteRequest || svelteRequest.query.type === 'style' || svelteRequest.raw) {
 					return;
 				}
 				let compileData;
@@ -245,11 +242,7 @@ export function svelte(inlineOptions?: Partial<Options>): Plugin[] {
 				}
 				const svelteRequest = requestParser(ctx.file, false, ctx.timestamp);
 				if (svelteRequest) {
-					try {
-						return handleHotUpdate(compileSvelte, ctx, svelteRequest, cache, options);
-					} catch (e) {
-						throw toRollupError(e, options);
-					}
+					return handleHotUpdate(compileSvelte, ctx, svelteRequest, cache, options);
 				}
 			},
 			async buildEnd() {
@@ -261,6 +254,7 @@ export function svelte(inlineOptions?: Partial<Options>): Plugin[] {
 	return plugins.filter(Boolean);
 }
 
+export { vitePreprocess } from './preprocess';
 export { loadSvelteConfig } from './utils/load-svelte-config';
 
 export {